Move Most Viewed Pages Macro to livetable #10 (#12)
I moved the implementation from JS to Java to utilize the livetable because the JS implementation would lead to CORS issues. For this move, I had to add the API and default modules, and create a page named JSONLivetableResults that will access the scripting service to make the request towards Matomo and normalize the response it gives, to have a specific format and alter some JSON nodes data. I also added Java configuration for the application and unit tests for all the public methods.